requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-oauth-toolkit>=2.2.0
requests>=2.28.0
pytz>=2023.3
pyyaml>=6.0
redis>=4.5.0